third_party/WebKit/LayoutTests/editing/pasteboard/drag-drop-copy-text-expected.txt

CONSOLE ERROR: line 9: Uncaught SyntaxError: Illegal return statement
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
